+++ packages/expo-linking/src/Linking.ts
@@ -264,6 +264,21 @@ export function addEventListener(type: 'url', handler: URLListener): EmitterSubs
   return NativeLinking.addEventListener(type, handler);
 }
 
+/**
+ * Remove a handler by passing the `url` event type and the handler.
+ * @param type The only valid type is `'url'`.
+ * @param handler An [`URLListener`](#urllistener) function that takes an `event` object of the type
+ * [`EventType`](#eventype).
+ * @see [React Native Docs Linking page](https://reactnative.dev/docs/linking#removeeventlistener).
+ *
+ * @deprecated Call `remove()` on the return value of `addEventListener()` instead.
+ */
+export function removeEventListener(type: 'url', handler: URLListener): void {
+  console.warn(
+    'The `removeEventListener()` method has been removed. Please call instead `remove()` on the return value of `addEventListener()`.'
+  );
+}
+
 // @needsAudit
 /**
  * Helper method which wraps React Native's `Linking.getInitialURL()` in `Linking.parse()`.